# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register a new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        },
                        "email": {
                          "type": "string",
                          "example": "user@example.com"
                        },
                        "fullName": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "role": {
                          "type": "string",
                          "example": "client"
                        }
                      }
                    },
                    "accessToken": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "email must be an email",
                        "password must be longer than or equal to 6 characters"
                      ]
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "User with this email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "example": "User with this email already exists"
                    },
                    "error": {
                      "type": "string",
                      "example": "Conflict"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Log in existing user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        },
                        "email": {
                          "type": "string",
                          "example": "user@example.com"
                        },
                        "fullName": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "role": {
                          "type": "string",
                          "example": "client"
                        }
                      }
                    },
                    "accessToken": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentials"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/consultants/{id}/verify": {
      "patch": {
        "operationId": "ConsultantsController_verify",
        "summary": "Verify a consultant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isVerified": {
                    "type": "boolean",
                    "description": "Verification status",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consultant verification updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consultant"
                }
              }
            }
          },
          "404": {
            "description": "Consultant not found"
          }
        },
        "tags": [
          "consultants"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/admin/dashboard": {
      "get": {
        "operationId": "AdminController_getDashboardStats",
        "summary": "Get dashboard statistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns dashboard statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStatsDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires admin role"
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/consultants/pending": {
      "get": {
        "operationId": "AdminController_getPendingConsultants",
        "summary": "Get list of pending consultants",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns consultants pending verification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Consultant"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires admin role"
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/consultants/{id}/verify": {
      "patch": {
        "operationId": "AdminController_verifyConsultant",
        "summary": "Verify or reject a consultant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "isVerified"
                ],
                "properties": {
                  "isVerified": {
                    "type": "boolean",
                    "description": "Verification status",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consultant verification status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consultant"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires admin role"
          },
          "404": {
            "description": "Consultant not found"
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Family Counseling API",
    "description": "\n      # Family Counseling API Documentation\n      \n      This API provides the backend services for the Family Counseling Platform.\n\n      ## Authorization\n      Most endpoints require JWT authentication. After logging in, include the JWT token in the Authorization header:\n      ```\n      Authorization: Bearer YOUR_JWT_TOKEN\n      ```\n      \n      ## Role-based access\n      Different endpoints have different role requirements. The API supports these roles:\n      - admin: Full access to all endpoints\n      - consultant: Access to consultant-specific features\n      - client: Access to client-specific features\n    ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "consultant",
              "client"
            ],
            "default": "client"
          },
          "phoneNumber": {
            "type": "string"
          }
        },
        "required": [
          "fullName",
          "email",
          "password",
          "role"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "Consultant": {
        "type": "object",
        "properties": {
          "specialties": {
            "description": "Consultant specialties",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bio": {
            "type": "string",
            "description": "Consultant biography"
          },
          "education": {
            "type": "string",
            "description": "Education and qualifications"
          },
          "rating": {
            "type": "number",
            "default": 0
          },
          "reviewCount": {
            "type": "number",
            "default": 0
          },
          "isVerified": {
            "type": "boolean",
            "default": false
          },
          "consultantLicense": {
            "type": "string"
          }
        },
        "required": [
          "specialties",
          "bio",
          "education",
          "rating",
          "reviewCount",
          "isVerified"
        ]
      },
...
      "DashboardStatsDto": {
        "type": "object",
        "properties": {
          "totalUsers": {
            "type": "number"
          },
          "totalConsultants": {
            "type": "number"
          },
          "totalClients": {
            "type": "number"
          },
          "totalSessions": {
            "type": "number"
          },
          "pendingSessions": {
            "type": "number"
          },
          "completedSessions": {
            "type": "number"
          },
          "totalReviews": {
            "type": "number"
          },
          "averageRating": {
            "type": "number"
          }
        },
        "required": [
          "totalUsers",
          "totalConsultants",
          "totalClients",
          "totalSessions",
          "pendingSessions",
          "completedSessions",
          "totalReviews",
          "averageRating"
        ]
      }
    }
  }
}